Fix 37 logic bugs in asynch, synch, sound, persistences, reflect, swing, knowledge and the root tools
Covers all 30 rows recorded for these folders in HANDOFF.md, plus seven
undocumented siblings of the same defects found alongside them:
- DirectPlayer.keyReleased had the identical unguarded NOTES_BY_KEY lookup
  as the documented keyPressed row, reachable via the same key.
- UniCastConstrained.removeValidator, isValidator and countValidators all
  carried the same dead 'instanceof MultiCaster' test as the documented
  addValidator row; MultiValidator does not extend MultiCaster, so the
  delegation never fired.
- EnumAttribute, MetricAttribute, StringAttribute and TimeAttribute each
  seeded Fields with an empty array instead of BasicAttribute.Keys, so the
  inherited TypeID/SubjectID/StatusID key columns were missing from every
  statement DBObjectFactory built from them, contradicting the
  'including Parent Fields' contract. MetaType shows the intended form.
- DirToXml's timestamp used 'hh' (12-hour, with no am/pm marker) as well
  as the documented month/minute swap, in an ISO-8601-shaped string.

Concurrency defects, all fixed conservatively within the existing locking
discipline:
- BlockedThreadExecutor notified the wrong monitor - the held lock is
  'this', not 't' - so the wakeup never arrived; its guard was an if
  rather than a while, and stop() was unsynchronized.
- ThreadPoolExecutor's worker polled and waited outside the monitor that
  ThreadExecutor.execute notifies on. The poll and wait now happen under
  it, with r.run() deliberately left outside so the pool stays concurrent.
- QueuedSemaphore.lock enqueued the thread before acquiring its own
  monitor, so an unlock() that dequeued it could reach notify() first and
  the waiter would sleep forever. The monitor is now taken before the
  enqueue and held until wait() releases it; the nesting is monitor then
  this, and unlock never takes this, so no lock-order cycle is introduced.
- Scheduler busy-spun; it now blocks on the Operations monitor,
  indefinitely when the queue is empty and for POLL_MILLIS when items
  remain dirty.
- ThreadExecutor decremented numTasks even when it took nothing.
- Barrier never initialized its countdown.

Other families:
- Field-instead-of-parameter: PersistedObject's constructor tested the
  null field ID rather than ID_, so setId never ran; reflect/Type had the
  same defect twice.
- sound: WaveStreamIn/WaveStreamOut read and wrote 4 bytes for a 24-bit
  sample, on both sides; WaveDataChunk read an int for a 16-bit sample;
  MidiChunk never filled its event buffer, leaving the stream unadvanced.
- knowledge/DBObjectFactory: substring(0, -1) on an empty field list, and
  column names taken from the Java field names rather than DBFieldNames
  and DBKeyNames.
- The root command-line tools fell through their usage messages instead of
  returning.

All 30 touched files compile under javac 11 with -encoding UTF-8 and keep
their original line endings. Verification is compile-only; the concurrency
fixes in particular are reasoned from the code, not demonstrated by a test.

Recorded while verifying: 41 of the 1457 .java files in this tree are
LF-only rather than CRLF, which predates this run.

+++ sound/WaveDataChunk.java
@@ -66,8 +66,7 @@ extends FileChunk {
 				stream16 = new int [chunkSize/numBytesPerValue];
 				stream_8 = null;
 				for (int i = -1; ++i < stream16.length; ) {
-					// TODO: LOGIC: readInt() consumes 4 Bytes per Sample, but a 16-Bit Sample is only 2 Bytes; this over-reads the Stream by 2x for the (most common) 16-Bit case, misaligning or exhausting the underlying Stream. Should use streamIn.readShort() here.
-					stream16[i] = streamIn.readInt(); }
+					stream16[i] = streamIn.readShort(); }
 				break;
 			case 3: //24 Bit
 				stream16 = null;
